Elide redundant bounds check

+++ src/Ramstack.FileProviders/Internal/FilePath.cs
@@ -123,7 +123,7 @@ internal static class FilePath
         {
             var prior = path[0];
 
-            for (var j = 1; j < path.Length; j++)
+            for (var j = 1; (uint)j < (uint)path.Length; j++)
             {
                 var ch = path[j];
                 if (ch == '\\' || ch == '/' && prior == '/')
